requirements.txt
pandas
cx_Oracle
#gitpython
schedule
click
pyhamcrest
pytest
pyyaml
git+ssh://git@github.com/brizjin/GitPython.git
git+ssh://git@gitlab.moduldev.ru/abs/database-client.git